requirements.txt
Django~=4.2.17
beautifulsoup4~=4.12.3
soupsieve~=2.6
pip~=24.3.1
asgiref~=3.8.1
sqlparse~=0.5.3
django-crispy-forms~=2.3
django-debug-toolbar~=4.4.6
django-bootstrap-v5 ~= 1.0.11
crispy-bootstrap5 ~= 2024.10
django-htmx~=1.21.0